added validation extensions
