+++ Halcyon.Web/Models/User/UpdateUserRequestValidator.cs
@@ -1,4 +1,5 @@
 ﻿using FluentValidation;
+using Halcyon.Web.Extensions;
 using Halcyon.Web.Services.Date;
 
 namespace Halcyon.Web.Models.User
@@ -10,7 +11,7 @@ namespace Halcyon.Web.Models.User
             RuleFor(x => x.EmailAddress).NotEmpty().EmailAddress().MaximumLength(255).WithName("Email Address");
             RuleFor(x => x.FirstName).NotEmpty().MaximumLength(50).WithName("First Name");
             RuleFor(x => x.LastName).NotEmpty().MaximumLength(50).WithName("Last Name");
-            RuleFor(x => x.DateOfBirth).NotEmpty().LessThan(dateService.UtcNow).WithName("Date Of Birth");
+            RuleFor(x => x.DateOfBirth).NotEmpty().InThePast(dateService).WithName("Date Of Birth");
         }
     }
 }
